+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # PayPal Android SDK Release Notes
 
-## unreleased
+## 0.0.8 (2023-03-13)
 * `PayPalNativePayments`:
   *  Bump `PayPal Native Checkout` to `0.8.8` and add `return_url`
 * Send analytic events for `PayPalNativePayments`, `PayPalWebPayments`, and `CardPayments` flows